change the method to GuessSchemaFields
