+++ server/internal/adapter/gql/gqlmodel/convert_asset.go
@@ -4,7 +4,6 @@ import (
 	"path/filepath"
 	"strings"
 
-	"github.com/reearth/reearth-cms/server/internal/usecase/interfaces"
 	"github.com/reearth/reearth-cms/server/pkg/asset"
 	"github.com/reearth/reearthx/usecasex"
 	"github.com/samber/lo"
@@ -143,27 +142,6 @@ func ToAssetFile(a *asset.File) *AssetFile {
 		Path:            a.Path(),
 		FilePaths:       a.FilePaths(),
 	}
-}
-
-func AssetsToSchemaFieldData(data *interfaces.AssetSchemaFieldsData) *AssetSchemaFieldResult {
-	// Return nil if input is nil
-	if data == nil {
-		return nil
-	}
-
-	result := &AssetSchemaFieldResult{
-		TotalCount: data.TotalCount,
-		Fields:     make([]*AssetSchemaField, 0, len(data.Fields)),
-	}
-
-	for _, f := range data.Fields {
-		result.Fields = append(result.Fields, &AssetSchemaField{
-			FieldName: f.FieldName,
-			FieldType: f.FieldType,
-		})
-	}
-
-	return result
 }
 
 func (s *AssetSort) Into() *usecasex.Sort {
