+++ server/internal/adapter/gql/gqlmodel/models_gen.go
@@ -527,6 +527,21 @@ type GroupsPayload struct {
 	Groups []*Group `json:"groups"`
 }
 
+type GuessSchemaField struct {
+	Key  string `json:"key"`
+	Type string `json:"type"`
+	Name string `json:"name"`
+}
+
+type GuessSchemaFieldResult struct {
+	TotalCount int                 `json:"total_count"`
+	Fields     []*GuessSchemaField `json:"fields"`
+}
+
+type GuessSchemaFieldsInput struct {
+	AssetID ID `json:"assetId"`
+}
+
 type Integration struct {
 	ID          ID                 `json:"id"`
 	Name        string             `json:"name"`
