+++ server/internal/adapter/gql/resolver_field.go
@@ -255,6 +255,19 @@ func (r *mutationResolver) DeleteField(ctx context.Context, input gqlmodel.Delet
 	}, nil
 }
 
+// GuessSchemaFields is the resolver for the guessSchemaFields field.
+func (r *queryResolver) GuessSchemaFields(ctx context.Context, input gqlmodel.GuessSchemaFieldsInput) (*gqlmodel.GuessSchemaFieldResult, error) {
+	id, err := id.AssetIDFrom(string(input.AssetID))
+	if err != nil {
+		return nil, err
+	}
+	fields, err := usecases(ctx).Schema.GuessSchemaFieldsByAssetID(ctx, id, getOperator(ctx))
+	if err != nil {
+		return nil, err
+	}
+	return gqlmodel.AssetsToGuessSchemaFieldData(fields), nil
+}
+
 // Model is the resolver for the model field.
 func (r *schemaFieldResolver) Model(ctx context.Context, obj *gqlmodel.SchemaField) (*gqlmodel.Model, error) {
 	if obj.ModelID != nil {
