+++ server/internal/usecase/interactor/schema.go
@@ -2,6 +2,10 @@ package interactor
 
 import (
 	"context"
+	"encoding/json"
+	"io"
+	"regexp"
+	"strings"
 
 	"github.com/reearth/reearth-cms/server/internal/usecase"
 	"github.com/reearth/reearth-cms/server/internal/usecase/gateway"
@@ -558,4 +562,162 @@ func (i Schema) CreateFields(ctx context.Context, sId id.SchemaID, createFieldsP
 
 			return s.Fields(), nil
 		})
+}
+
+// GeoJSONFeature represents a basic GeoJSON structure
+type GeoJSONFeature struct {
+	Type       string                 `json:"type"`
+	Geometry   map[string]interface{} `json:"geometry"`
+	Properties map[string]interface{} `json:"properties"`
+}
+
+func (i *Schema) GuessSchemaFieldsByAssetID(ctx context.Context, assetID id.AssetID, operator *usecase.Operator) (*interfaces.GuessSchemaFieldsData, error) {
+	if operator.AcOperator.User == nil && operator.Integration == nil {
+		return &interfaces.GuessSchemaFieldsData{}, interfaces.ErrInvalidOperator
+	}
+
+	asset, err := i.repos.Asset.FindByID(ctx, assetID)
+	if err != nil {
+		return &interfaces.GuessSchemaFieldsData{}, err
+	}
+
+	assetFileData, err := i.repos.AssetFile.FindByID(ctx, assetID)
+	if err != nil {
+		return &interfaces.GuessSchemaFieldsData{}, err
+	}
+
+	// check if file is a json or geojson
+	if assetFileData.ContentType() != "application/json" && assetFileData.ContentType() != "application/geo+json" {
+		return &interfaces.GuessSchemaFieldsData{}, interfaces.ErrInvalidContentTypeForSchemaConversion
+	}
+
+	// read file
+	file, _, err := i.gateways.File.ReadAsset(ctx, asset.UUID(), assetFileData.Path(), nil)
+	if err != nil {
+		return &interfaces.GuessSchemaFieldsData{}, err
+	}
+	// read file content
+	content, err := io.ReadAll(file)
+	if err != nil {
+		return &interfaces.GuessSchemaFieldsData{}, err
+	}
+
+	schemaFieldData, err := extractSchemaFieldData(assetFileData.ContentType(), content)
+	if err != nil {
+		return &interfaces.GuessSchemaFieldsData{}, err
+	}
+	return schemaFieldData, nil
+}
+
+func extractSchemaFieldData(contentType string, content []byte) (*interfaces.GuessSchemaFieldsData, error) {
+	result := &interfaces.GuessSchemaFieldsData{}
+
+	switch contentType {
+	case "application/json":
+		var raw map[string]any
+		if err := json.Unmarshal(content, &raw); err != nil {
+			return result, err
+		}
+
+		// ✅ Detect JSON Schema
+		if isJSONSchema(raw) {
+			// Extract from "properties"
+			if props, ok := raw["properties"].(map[string]any); ok {
+				for name, val := range props {
+					if def, ok := val.(map[string]any); ok {
+						typeValue, ok := def["type"].(string)
+						if !ok {
+							continue
+						}
+						result.Fields = append(result.Fields, interfaces.GuessSchemaField{
+							Name: name,
+							Type: typeValue,
+							Key:  toKey(name),
+						})
+					}
+				}
+				result.TotalCount = len(result.Fields)
+				return result, nil
+			} else {
+				return result, interfaces.ErrInvalidJSONSchema
+			}
+		} else {
+			return result, interfaces.ErrInvalidJSONSchema
+		}
+
+	case "application/geo+json":
+		var geoData GeoJSONFeature
+		if err := json.Unmarshal(content, &geoData); err != nil {
+			return result, err
+		}
+
+		if geoData.Properties == nil {
+			return result, nil
+		}
+
+		for name, value := range geoData.Properties {
+			result.Fields = append(result.Fields, interfaces.GuessSchemaField{
+				Name: name,
+				Type: detectFieldType(value),
+				Key:  toKey(name),
+			})
+		}
+	}
+
+	result.TotalCount = len(result.Fields)
+	return result, nil
+}
+
+func isJSONSchema(data map[string]interface{}) bool {
+	_, hasSchema := data["$schema"]
+	_, hasProperties := data["properties"]
+	typeVal, isTypeObj := data["type"].(string)
+	return hasSchema && hasProperties && isTypeObj && typeVal == "object"
+}
+
+func detectFieldType(value any) string {
+	switch v := value.(type) {
+	case string:
+		return "string"
+	case float64:
+		// If the number is whole, assume it's an integer
+		if v == float64(int(v)) {
+			return "integer"
+		}
+		return "float"
+	case bool:
+		return "boolean"
+	case map[string]interface{}:
+		return "object"
+	case []interface{}:
+		return "array"
+	default:
+		return "unknown"
+	}
+}
+
+// toKey converts a title string into a valid key:
+// - Lowercases everything
+// - Replaces spaces with hyphens
+// - Removes invalid characters (only allows a-z, 0-9, -, _)
+// - Truncates to 32 characters max
+func toKey(title string) string {
+	// Lowercase
+	key := strings.ToLower(title)
+
+	// Replace spaces and consecutive whitespace with hyphens
+	key = regexp.MustCompile(`\s+`).ReplaceAllString(key, "-")
+
+	// Remove all characters that are not a-z, 0-9, -, _
+	key = regexp.MustCompile(`[^a-z0-9_-]`).ReplaceAllString(key, "")
+
+	// Trim to max length of 32
+	if len(key) > 32 {
+		key = key[:32]
+	}
+
+	// Remove leading/trailing hyphens (optional cleanup)
+	key = strings.Trim(key, "-")
+
+	return key
 }
